requirements.txt
playwright
python-dotenv
openpyxl
